commiting my stuff-had a slight error doing it
had an error commiting this. For some reason my code only appeared as
chinese characters on the git preview. I had to slowly copy everything
line by line into a txt document for it to not spaz and becoem chinese.
Hopefully it stops, but i also commited the chinese document to see if
anyone can help, ill look online ans ask the group if the problem
continues.